requirements.txt.txt
python-telegram-bot==20.7
pillow>=10.0.0
pymupdf>=1.23.0